Modified Dockerfile to install MS-SQL command-line tools.
At this time, Microsoft's package is compatible with Ubuntu 16.04
but not 18.04. Therefore, the Dockerfile won't build as is.
The changes here are mostly stolen from Microsoft's own Dockerfile,
which is hosted at https://github.com/Microsoft/mssql-docker and will
presumably be updated when 18.04 is supported. The Docker image is
hosted at https://hub.docker.com/_/microsoft-mssql-tools.